+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 ### Changed
 
 * `Checkbox` is slightly smaller in size
+* Default `Button` type to `button`
+* Respect semantic `disabled` attribute on `Button` when disabled
 * Bumped `react` peer dependency version to `^16.3.1`
 
 ## v1.1.2